// src/glossary/GlossaryProps.d.ts
export declare type Term = string;
/**
 * Select an image to display inside the content box, at the top
 */
export declare type Definition = string;
/**
 * Url to main image for this glossary entry
 */
export declare type UrlForTheMainImage = string;
/**
 * Title of the term being defined
 */
export declare type CaptionForTheMainImage = string;
/**
 * Select a media asset for this glossary entry
 */
export declare type UrlForTheMediaAsset = string;
/**
 * Title of the term being defined
 */
export declare type CaptionForTheMainImage1 = string;
/**
 * Media assets to display for this glossary entry
 */
export declare type Media = MediaAsset[];
/**
 * Name of the tag
 */
export declare type Tag = string;
/**
 * Tags for this glossary entry
 */
export declare type Tags = Tag[];
/**
 * Title of the related term
 */
export declare type Title = string;
/**
 * Excerpt for the related term
 */
export declare type Excerpt = string;
/**
 * Url for the related term
 */
export declare type Url = string;
/**
 * Image for the related term
 */
export declare type Image = string;
/**
 * Entries related to this glossary entry
 */
export declare type RelatedEntries = RelatedEntry[];
/**
 * Link to the related StacKShare discussion
 */
export declare type StackShareUrl = string;
// ...
export interface GlossaryProps {
    term: Term;
    definition?: Definition;
    cover?: CoverImage;
    media?: Media;
    tags?: Tags;
    related?: RelatedEntries;
    stackshare?: StackShareUrl;
    cta?: Cta;
    [k: string]: unknown;
}
/**
 * Add main image to represent this glossary entry
 */
export interface CoverImage {
    src?: UrlForTheMainImage;
    caption?: CaptionForTheMainImage;
    [k: string]: unknown;
}
/**
 * Single media asset
 */
export interface MediaAsset {
    src?: UrlForTheMediaAsset;
    caption?: CaptionForTheMainImage1;
    [k: string]: unknown;
}
/**
 * Single related entry
 */
export interface RelatedEntry {
    title?: Title;
    excerpt?: Excerpt;
    url?: Url;
    image?: Image;
    [k: string]: unknown;
}
export interface Cta {
    headline: Headline;
    storytelling: Storytelling;
    button: LinkButton1;
    [k: string]: unknown;
}
/**
 * Headline
 */
export interface Headline {
    align: Alignment;
    level: Level;
    styleAs?: Style;
    content?: Text;
    subheadline?: Subheadline;
    spaceAfter: BottomSpacing;
    pageHeader?: PageHeader;
    className?: AdditionalClasses;
    [k: string]: unknown;
}
/**
 * Component to present rich combinations of text and media, best used sequentially
 */
export interface Storytelling {
    image: Image1;
    backgroundImage?: BackgroundImage;
    backgroundColor?: BackgroundColor;
    full?: FullSizeImage;
    box: TextBox;
    className?: Class;
    [k: string]: unknown;
}
// ...
export interface LinkButton1 {
    label: Label1;
    variant: ButtonStyle1;
    size: ButtonSize1;
    className?: AdditionalClasses3;
    icon?: Icon1;
    iconBefore?: IconBeforeButton1;
    iconAfter?: IconAfterButton1;
    dataComponent?: DataComponentAttribute1;
    fillAnimation?: FillAnimation1;
    iconAnimation?: IconAnimation1;
    href: ButtonHref1;
    newTab?: OpenLinkInNewTab1;
}